bw.taylor {spatstat.univar}	R Documentation

Bandwidth Selection for Kernel Density Estimation by Non-Random Bootstrap

Description

Use Taylor's non-random bootstrap technique to select the bandwidth for kernel density estimation on the real line.

Usage

bw.taylor(x, ..., srange = NULL, useC = TRUE)

Arguments

x
Numeric vector.

...
Ignored.

srange
Range of bandwidths to be considered. A numeric vector of length 2.

useC
Logical value specifying whether to use faster C code.

Details

This function selects a bandwidth for kernel density estimation of a probability density on the real line, using the numeric data x and assuming a Gaussian kernel. The result is the numeric value of the standard deviation of the Gaussian kernel.

The function uses the method of Taylor (1989) who showed that, when using the Gaussian kernel, the optimisation criterion can be computed rapidly from the data without any randomised resampling.

The domain of the probability density is assumed to be the entire real line. Boundary correction is not currently implemented.

The result of bw.taylor is a single numeric value giving the selected bandwidth.

Value

A single numeric value.

Author(s)

Tilman Davies Tilman.Davies@otago.ac.nz and Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

References

Taylor, C.C. (1989) Choice of the Smoothing Parameter in Kernel Density Estimation, Biometrika 76 4, 705–712.

See Also

bw.nrd in the stats package for standard bandwidth selectors.

Examples

  x <- rnorm(30)
  bw.taylor(x)

[Package spatstat.univar version 3.1-1 Index]