densityAdaptiveKernel {spatstat.univar}	R Documentation

Adaptive Kernel Estimation of Density or Intensity

Description

Computes an adaptive estimate of probability density or intensity using a variable-bandwidth smoothing kernel.

Usage

  densityAdaptiveKernel(X, ...)

Arguments

X
Data to be smoothed.

...
Additional arguments passed to methods.

Details

This generic function computes an adaptive kernel estimate of probability density or intensity.

The function densityAdaptiveKernel is generic. The spatstat package family includes methods for spatial objects.

Value

See documentation for each method.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Tilman Davies Tilman.Davies@otago.ac.nz.

See Also

bw.abram.

[Package spatstat.univar version 3.1-1 Index]