integral {spatstat.univar}	R Documentation

Integral of a Function or Spatial Object

Description

Computes the integral of a function or spatial object.

Usage

integral(f, domain=NULL, ...)

Arguments

f
A function, or a spatial object that can be treated as a function.

domain
Optional. Data specifying the domain of integration.

...
Arguments passed to methods.

Details

The function integral is generic. It calculates the integral of a function, or the integral of a spatial object that can be treated as a function. It has methods for one-dimensional functions ("density", "fv") and for spatial objects ("im", "msr", "linim", "linfun").

Value

A single numeric or complex value, or a vector of such values.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

integral.density.

integral.im in package spatstat.geom.

[Package spatstat.univar version 3.1-1 Index]