kernel.factor {spatstat.univar}	R Documentation

Scale factor for density kernel

Description

Returns a scale factor for the kernels used in density estimation for numerical data.

Usage

  kernel.factor(kernel = "gaussian")

Arguments

kernel
String name of the kernel. Options are "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine" and "optcosine". (Partial matching is used).

Details

Kernel estimation of a probability density in one dimension is performed by density.default using a kernel function selected from the list above.

This function computes a scale constant for the kernel. For the Gaussian kernel, this constant is equal to 1. Otherwise, the constant c is such that the kernel with standard deviation 1 is supported on the interval [-c,c].

For more information about these kernels, see density.default.

Value

A single number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Martin Hazelton Martin.Hazelton@otago.ac.nz.

See Also

density.default, dkernel, kernel.moment, kernel.squint

Examples

   kernel.factor("rect")
   # bandwidth for Epanechnikov kernel with half-width h=1
   h <- 1
   bw <- h/kernel.factor("epa")

[Package spatstat.univar version 3.1-1 Index]