kernel.squint {spatstat.univar}	R Documentation

Integral of Squared Kernel

Description

Computes the integral of the squared kernel, for the kernels used in density estimation for numerical data.

Usage

  kernel.squint(kernel = "gaussian", bw=1)

Arguments

kernel
String name of the kernel. Options are "gaussian", "rectangular", "triangular", "epanechnikov", "biweight", "cosine" and "optcosine". (Partial matching is used).

bw
Bandwidth (standard deviation) of the kernel.

Details

Kernel estimation of a probability density in one dimension is performed by density.default using a kernel function selected from the list above.

This function computes the integral of the squared kernel,

R = integral of k(x)^2 dx from x = -infinity to x = +infinity

where k(x) is the kernel with bandwidth bw.

Value

A single number.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au and Martin Hazelton Martin.Hazelton@otago.ac.nz.

See Also

density.default, dkernel, kernel.moment, kernel.factor

Examples

   kernel.squint("gaussian", 3)

   # integral of squared Epanechnikov kernel with half-width h=1
   h <- 1
   bw <- h/kernel.factor("epa")
   kernel.squint("epa", bw)

[Package spatstat.univar version 3.1-1 Index]