quantile.density {spatstat.univar}	R Documentation

Quantiles of a Density Estimate

Description

Given a kernel estimate of a probability density, compute quantiles.

Usage

## S3 method for class 'density'
quantile(x, probs = seq(0, 1, 0.25), names = TRUE,
        ..., warn = TRUE)

Arguments

x
Object of class "density" computed by a method for density

probs
Numeric vector of probabilities for which the quantiles are required.

names
Logical value indicating whether to attach names (based on probs) to the result.

...
Ignored.

warn
Logical value indicating whether to issue a warning if the density estimate x had to be renormalised because it was computed in a restricted interval.

Details

This function calculates quantiles of the probability distribution whose probability density has been estimated and stored in the object x. The object x must belong to the class "density", and would typically have been obtained from a call to the function density.

The probability density is first normalised so that the total probability is equal to 1. A warning is issued if the density estimate was restricted to an interval (i.e. if x was created by a call to density which included either of the arguments from and to).

Next, the density estimate is numerically integrated to obtain an estimate of the cumulative distribution function F(x). Then for each desired probability p, the algorithm finds the corresponding quantile q.

The quantile q corresponding to probability p satisfies F(q) = p up to the resolution of the grid of values contained in x. The quantile is computed from the right, that is, q is the smallest available value of x such that F(x) >= p.

Value

A numeric vector containing the quantiles.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

quantile, quantile.ewcdf, CDF.

Examples

   dd <- density(runif(10))
   quantile(dd)

[Package spatstat.univar version 3.1-1 Index]