stieltjes {spatstat.univar}	R Documentation

Compute Integral of Function Against Cumulative Distribution

Description

Computes the Stieltjes integral of a function f with respect to a function M.

Usage

stieltjes(f, M, ...)

Arguments

f
The integrand. A function in the R language.

M
The cumulative function against which f will be integrated. An object of class "fv" or "stepfun".

...
Additional arguments passed to f.

Details

This command computes the Stieltjes integral

I = integral f(x) dM(x)

of a real-valued function f(x) with respect to a nondecreasing function M(x).

One common use of the Stieltjes integral is to find the mean value of a random variable from its cumulative distribution function F(x). The mean value is the Stieltjes integral of f(x)=x with respect to F(x).

The argument f should be a function in the R language. It should accept a numeric vector argument x and should return a numeric vector of the same length.

The argument M should be either a step function (object of class "stepfun") or a function value table (object of class "fv" ). Objects of class "stepfun" are returned by ecdf, ewcdf, and other utilities.

Value

A list containing the value of the Stieltjes integral computed using each of the versions of the function M.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

Examples

  x <- runif(100)
  w <- runif(100)
  H <- ewcdf(x, w)
  stieltjes(function(x) { x^2 }, H)

[Package spatstat.univar version 3.1-1 Index]