knots.ewcdf {spatstat.univar}	R Documentation

Jump Points of an Empirical Weighted Cumulative Distribution Function

Description

Extract the knots (jump points) of an empirical weighted cumulative distribution function.

Usage

 ## S3 method for class 'ewcdf'
knots(Fn, ...)

Arguments

Fn
An empirical weighted cumulative distribution function (object of class "ewcdf").

...
Ignored.

Details

The function knots is generic. This function knots.ewcdf is the method for the class "ewcdf" of empirical weighted cumulative distribution functions. Objects of class "ewcdf" are created by ewcdf.

The jump points (locations of increments) of the function will be returned as a numeric vector.

Value

Numeric vector.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

See Also

ewcdf, quantile.ewcdf

Examples

   x <- c(1, 2, 5)
   w <- runif(3)
   e <- ewcdf(x,w)
   knots(e)

[Package spatstat.univar version 3.1-1 Index]